moved board.go from core/game to core/board
