+++ core/board/board_test.go
@@ -1,4 +1,4 @@
-package game
+package board
 
 import (
 	"fmt"
